requirements.txt
python-telegram-bot == 20.3
qrcode[pil] == 7.4.2